requirement.txt
mysql-connector-python==8.0.23
prettytable==2.1.0
protobuf==3.15.6
six==1.15.0
wcwidth==0.2.5
